v2.42.671 BETA DEV
- Managed to get the TileEntities to work!
- Still having issues with Slots
- Fixed BrainLightSensor (including the smoking effect)
- Please test this build but DO NOT use it on your old worlds. It might
corrupt them!

+++ main/java/brainstonemod/network/packet/BrainStoneUpdateTileEntityPacket.java
@@ -34,7 +34,7 @@ public class BrainStoneUpdateTileEntityPacket extends BrainStoneBasePacket {
 			final PacketBuffer buffer = new PacketBuffer(data);
 
 			buffer.writeInt(x);
-			buffer.writeInt(y);
+			buffer.writeShort(y);
 			buffer.writeInt(z);
 			buffer.writeNBTTagCompoundToBuffer(nbt);
 		} catch (final IOException e) {
